Stack and release color names memory
Don't call the color_names memory freeing code explicitly in the
cleanup function as it is taken care of via {STACK,RELEASE}_VAR().
This implies the memory is "garbage collected" like other uses of
those macros, too.

Also, declare and initialize the color_names array in a non-global,
tighter scope.